refactor(app): 升级 AGP 9 并清理冗余告警

+++ app/src/main/java/com/wenhao/record/tracking/BackgroundTrackingService.kt
@@ -803,7 +803,7 @@ class BackgroundTrackingService : Service() {
                             endTimestamp = segment.endTimestamp,
                             segmentType = segment.kind.name,
                             confidence = defaultSegmentConfidence(segment.kind),
-                            distanceMeters = distanceMeters.toFloat(),
+                            distanceMeters = distanceMeters,
                             durationMillis = durationMillis,
                             avgSpeedMetersPerSecond = avgSpeedMetersPerSecond.toFloat(),
                             maxSpeedMetersPerSecond = maxSpeedMetersPerSecond.toFloat(),
